requirements.txt
ollama
rich
typer
pyyaml
